refactor(racing-car): separate functions, classes
- 함수 분리
- 클래스 분리
