+++ docs/README.md
@@ -51,8 +51,8 @@
 
 ### 추가사항
 - [ ] 코드 리팩토링을 진행한다.
-    - [ ] 함수를 분리한다.
-    - [ ] 클래스를 분리한다.
+    - [x] 함수를 분리한다.
+    - [x] 클래스를 분리한다.
 - [ ] 주석을 Javadoc 형태로 변경한다.
 - [ ] README 문서를 상세하게 작성한다.
