+++ src/main/java/racingcar/Application.java
@@ -1,6 +1,6 @@
 package racingcar;
 
-import racingcar.controller.CarRacingController;
+import racingcar.domain.CarRacing;
 
 import java.util.Scanner;
 
@@ -8,6 +8,6 @@ public class Application {
     public static void main(String[] args) {
         final Scanner scanner = new Scanner(System.in);
         // TODO 구현 진행
-        CarRacingController.startCarRacing(scanner);
+        CarRacing.runCarRacing(scanner);
     }
 }
