+++ src/main/java/racingcar/view/InputView.java
@@ -1,6 +1,6 @@
 package racingcar.view;
 
-import racingcar.Car;
+import racingcar.domain.Car;
 import racingcar.type.ErrorType;
 import racingcar.type.TextType;
 import utils.InputUtils;
@@ -13,10 +13,13 @@ import java.util.*;
  * @author Daeun Lee
  */
 public class InputView {
-
-    public static ArrayList<Car> scanCars(ArrayList<Car> cars, Scanner scanner) {
+    public static ArrayList<Car> scanCars(Scanner scanner) {
         String scannerCars = scanner.nextLine();
-        List<String> carNames = new ArrayList<>(Arrays.asList(scannerCars.split(TextType.COMMA_TEXT.getText())));
+
+        List<String> carNames = new ArrayList<>(
+                Arrays.asList(scannerCars.split(TextType.COMMA_TEXT.getText())));
+        ArrayList<Car> cars = new ArrayList<>();
+
         for (String carName : carNames) {
             cars.add(InputUtils.validateCar(carName));
         }
@@ -31,6 +34,7 @@ public class InputView {
             System.out.println(ErrorType.INVALID_CHARACTER.getError());
         }
         int scannerTimes = scanner.nextInt();
+
         InputUtils.validateTimes(scannerTimes);
 
         return scannerTimes;
